+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - Generic parser: added support for non-GBP exports.
 - Trezor Suite parser: added support for non-GBP exports.
 - Uphold parser: added support for non-GBP exports.
+- Kraken parser: added new export format. ([#355](https://github.com/BittyTax/BittyTax/issues/355))
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))